typo fix, checked_operations in other file
